Added tests for many common time domain analyses
